Old CDN service will be terminated on January 30
Due to China ICP register issues, we will no longer be able to provide CDN services from Chinese service providers. We've changed our CDN service to jsDelivr from v2.2.0, and the previous CDN service will be terminated on Jan 30, 2020. Please replace your old CDN to the new link in the README.md so as not to affect your business. We apologize for the inconvenience.